removed some warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@3534 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSVehicleType.cpp
@@ -128,7 +128,7 @@ MSVehicleType*
 MSVehicleType::dict_Random()
 {
     assert(myDict.size()!=0);
-    size_t r = randSUMO() * myDict.size();
+    size_t r = (size_t) (randSUMO() * myDict.size());
     if (r>=myDict.size()) {
         r = myDict.size() - 1;
     }
